♻️🏢 Split edit FF route
