+++ apps/admin/src/routes/_layout.foster-families.$id.edit/route-params.ts
@@ -1,0 +1,5 @@
+import { zu } from "@animeaux/zod-utils";
+
+export const routeParamsSchema = zu.object({
+  id: zu.string().uuid(),
+});
